nueva actualizacion

+++ Frontend/src/routes/dashboard.tsx
@@ -1,17 +1,26 @@
 import React, { useEffect, useState } from 'react';
-import {PortalLayout} from "../layout/PortalLayout";
+import { PortalLayout } from "../layout/PortalLayout";
 import { useAuth } from "../Autentication/AutProvider";
 import "./dashboard.css";
 import { useRef } from 'react';
+import Modal from 'react-modal'
+import '../../node_modules/bootstrap/dist/css/bootstrap.min.css';
+
 type Publicacion = {
   image: string;
   description: string;
-  _id:string;
+  _id: string;
+};
+type allUserPost = {
+  image: string;
+  description: string;
+  _id: string;
+  name: string;
 };
 
 export default function Dashboard() {
   const auth = useAuth();
-  
+
   const [, setEditingProfileImage] = useState(false);
   const [downloadURL, setDownloadURL] = useState("");
   const [description, setDescription] = useState("");
@@ -19,21 +28,47 @@ export default function Dashboard() {
   const [error, setError] = useState("");
   const fileInputRef = useRef<HTMLInputElement>(null); // Referencia para el input de tipo file
   const descriptionTextareaRef = useRef<HTMLTextAreaElement>(null);
-  
+  const [modalIsOpen, setModalIsOpen] = useState(false)
+  const [currenImage, setCurrenImage] = useState(null)
+  const [, setCurrentPublicationId] = useState(null)
+  const [publicacionesUsuarios, setPublicacionesUsuarios] = useState<allUserPost[]>([]);
+  const [isVisible, setIsVisible] = useState(false);
+
+  useEffect(() => {
+    // Ocultar el ul al inicio de la aplicación
+    setIsVisible(false);
 
 
-   
+    // Lógica para obtener las publicaciones de todos los usuarios
+    getPublishAllUsers();
+  }, []);
+
 
   useEffect(() => {
     getImageProfile()// Llama a la función para obtener la imagen de perfil cuando el componente se monta
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []); // El arreglo de dependencias vacío garantiza que el efecto se ejecute solo una vez
   useEffect(() => {
+    Modal.setAppElement('body')
     obtenerTodasLasPublicaciones()// Llama a la función para obtener la imagen de perfil cuando el componente se monta
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [])
 
-  
+  const handleFormSubmit = async (event) => {
+    event.preventDefault();
+    // Agrega aquí la lógica para obtener todas las publicaciones
+    // Puedes llamar a la función que obtiene las publicaciones aquí
+    // Por ejemplo: await getPublishAllUsers();
+    setIsVisible(false); // Muestra la lista de publicaciones después de obtenerlas
+  };
+
+  const handleFormSubmitFalse = async (event) => {
+    event.preventDefault();
+    // Agrega aquí la lógica para obtener todas las publicaciones
+    // Puedes llamar a la función que obtiene las publicaciones aquí
+    // Por ejemplo: await getPublishAllUsers();
+    setIsVisible(true); // Muestra la lista de publicaciones después de obtenerlas
+  };
 
   async function handleProfileImageChange(files: FileList | null) {
     try {
@@ -46,7 +81,7 @@ export default function Dashboard() {
       const formData = new FormData();
       formData.append("file", file);
       console.log('post TOKEN ', auth.getAccessToken());
-      
+
 
       // Subir la nueva imagen
       const responsePost = await fetch("http://localhost:5000/api/upload", {
@@ -56,8 +91,8 @@ export default function Dashboard() {
         },
         body: formData,
       });
-       console.log('POST perfil ',responsePost);
-       
+      console.log('POST perfil ', responsePost);
+
       if (!responsePost.ok) {
         console.error(
           "Error al cambiar la imagen de perfil:",
@@ -79,9 +114,9 @@ export default function Dashboard() {
   async function getImageProfile() {
     try {
       console.log('este es el token => ', auth.getAccessToken());
-      const id=  auth.getUser()?.id
-      
-      console.log('ID of user Profile',id);
+      const id = auth.getUser()?.id
+
+      console.log('ID of user Profile', id);
       const response = await fetch(`http://localhost:5000/api/getImage/${id}`, {
         method: 'GET',
         headers: {
@@ -89,8 +124,8 @@ export default function Dashboard() {
           Authorization: `Bearer ${auth.getAccessToken()}`,
         },
       });
-         console.log('GET perfil',response);
-         
+      console.log('GET perfil', response);
+
       if (!response.ok) {
         throw new Error('Error al obtener la imagen de perfil: ' + response.statusText);
       }
@@ -105,39 +140,42 @@ export default function Dashboard() {
       setDownloadURL(data.imageProfile); // Cambio aquí de data.downloadURL a data.imageProfile
       console.log('URL de la imagen de perfil obtenida:', data.imageProfile); // Cambio aquí de data.downloadURL a data.imageProfile
     } catch (error) {
+      const defaultImageUrl = 'https://static.vecteezy.com/system/resources/previews/003/337/584/large_2x/default-avatar-photo-placeholder-profile-icon-vector.jpg'; // Reemplaza con la URL de tu imagen por defecto
+      setDownloadURL(defaultImageUrl)
       console.error('Error al obtener la imagen de perfil:', error);
     }
     //
   }
+  // AGREGAR PUBLICACION
 
   async function postPublication(event: React.FormEvent<HTMLFormElement>) {
     event.preventDefault();
-  
+
     try {
       const formData = new FormData();
-  
+
       // Obtener el archivo seleccionado del input de tipo file
       const files = fileInputRef.current?.files;
-  
+
       // Obtener el valor del textarea
       const description = descriptionTextareaRef.current?.value;
-  
+
       // Agregar la descripción al FormData
       formData.append('description', description || '');
-  
+
       if (!files || files.length === 0) {
         console.error('No se han seleccionado archivos.');
         return;
       }
-  
+
       // Agregar el archivo al FormData
       for (let i = 0; i < files.length; i++) {
         formData.append('file', files[i]);
       }
-  
+
       const accessToken = auth.getAccessToken();
       console.log('Token de publicacion', accessToken);
-  
+
       const response = await fetch("http://localhost:5000/api/publicationpost", {
         method: "POST",
         headers: {
@@ -145,41 +183,42 @@ export default function Dashboard() {
         },
         body: formData,
       });
-  
+
       console.log('POST publicacion:', response);
-  
+
       if (!response.ok) {
         console.error("Error al cargar la imagen de publicación:", response.statusText);
         return;
       }
-  
+
       const data = await response.json();
       console.log("Respuesta del servidor al subir publicacion POST:", data);
-  
+
       // Actualizar la lista de publicaciones después de la publicación exitosa
       obtenerTodasLasPublicaciones();
-  
+      getPublishAllUsers();
+
       // Realizar cualquier acción necesaria después de cargar la nueva publicación
       // setFile(data.file);
       fileInputRef.current.value = ''; // Limpiar el input de tipo file
       setDescription('');
       console.log('url de imagen publicada',);
-  
+
     } catch (error) {
       console.error("Error al cargar la imagen publicada:", error);
     }
   }
-  
+  //OBTENER PUBLICACION
   const obtenerTodasLasPublicaciones = async () => {
     try {
       const accessToken = auth.getAccessToken();
       console.log('token de obtener la publicada ', accessToken);
-     console.log('elemetos de user',auth.getUser());
-    
-     
-     
-      const id= auth.getUser()?.id
-      console.log('ID of user',id);
+      console.log('elemetos de user', auth.getUser());
+
+
+
+      const id = auth.getUser()?.id
+      console.log('ID of user', id);
       const response = await fetch(`http://localhost:5000/api/publicationget/${id}`, {
         method: 'GET',
         headers: {
@@ -188,20 +227,20 @@ export default function Dashboard() {
         },
       });
       console.log('GET publicaciones', response);
-  
+
       if (!response.ok) {
         throw new Error('Error al obtener todas las publicaciones');
       }
       const data = await response.json();
-  
+
       // Ordenar las publicaciones por fecha de forma descendente
       const publicacionesOrdenadas = data.publications.slice().reverse();
 
-      console.log( 'estas son las puplicaiones en el orden que se debe revisar',publicacionesOrdenadas);
-      
+      console.log('estas son las puplicaiones en el orden que se debe revisar', publicacionesOrdenadas);
+
       setPublicaciones(publicacionesOrdenadas);
       console.log('este es data de la publicacion traida ', publicacionesOrdenadas);
-  
+
     } catch (error) {
       if (typeof error === 'string') {
         setError(error);
@@ -210,36 +249,82 @@ export default function Dashboard() {
       }
     }
   };
-  async function handleDeletePublication(publicationId:string) {
-    try {
-       const idPublicacion= auth.getUser.publicacion
-      console.log('ID de la publicación a eliminar:', publicationId);
-      
 
-      const accessToken = auth.getAccessToken();
-      const response = await fetch(`http://localhost:5000/api/delete/${auth.getUser()?.id}/publications/${publicationId}`, {
+  const modalHandler = (isOpen, image, publicationId) => {
+    setModalIsOpen(isOpen)
+    setCurrenImage(image)
+    setCurrentPublicationId(publicationId);
+  }
+  const deleteHandler = async () => {
+    try {
+      // Obtener los IDs de las imágenes de las publicaciones del usuario
+      const imageIDs = auth.getUser()?.publication.map(pub => pub._id);
+
+      // Validar si existe la imagen actual en las publicaciones del usuario
+      if (!imageIDs || !currenImage) {
+        console.error('No se puede eliminar la imagen actual porque no se ha seleccionado una imagen válida.');
+        return;
+      }
+
+      // Encontrar el ID de la imagen actual en las publicaciones del usuario
+      const imageIDToDelete = imageIDs.find(id => id === currenImage);
+
+      if (!imageIDToDelete) {
+        console.error('No se encontró el ID de la imagen actual en las publicaciones del usuario.');
+        return;
+      }
+
+      // Realizar la eliminación de la imagen utilizando su ID
+      const idUser = auth.getUser()?.id;
+      const response = await fetch(`http://localhost:5000/api/delete/${idUser}/publications/${imageIDToDelete}`, {
         method: 'DELETE',
         headers: {
           'Content-Type': 'application/json',
-          Authorization: `Bearer ${accessToken}`,
+          Authorization: `Bearer ${auth.getAccessToken()}`,
         },
       });
-       console.log('metodo eliminar',response);
-       
+
       if (!response.ok) {
-        throw new Error('Error al eliminar la publicación');
+        throw new Error('Error al eliminar la imagen.');
       }
-  
-      // Actualizar la lista de publicaciones después de eliminar
-      obtenerTodasLasPublicaciones();
+
+      console.log('Imagen eliminada exitosamente.');
+      // Aquí puedes realizar cualquier actualización de estado o acción necesaria después de la eliminación
+
     } catch (error) {
-      console.error('Error al eliminar la publicación:', error);
+      console.error('Error al eliminar la imagen:', error);
     }
-  }
-  
-  
+  };
+  const getPublishAllUsers = async () => {
+    try {
+      const response = await fetch(`http://localhost:5000/api/publicationgetAll`, {
+        method: 'GET',
+        headers: {
+          'Content-Type': 'application/json'
+        },
+      });
+      console.log('GET publicaciones', response);
 
+      if (!response.ok) {
+        throw new Error('Error al obtener todas las publicaciones all');
+      }
+      const data = await response.json();
 
+      // Ordenar las publicaciones por fecha de forma descendente
+      const publicacionesOrdenadas = data.publications.slice().reverse();
+
+      console.log('estas son las puplicaiones en el orden que se debe revisar all', publicacionesOrdenadas);
+
+      setPublicacionesUsuarios(publicacionesOrdenadas);
+      console.log('este es data de la publicacion traida all ', publicacionesOrdenadas);
+    } catch (error) {
+      if (typeof error === 'string') {
+        setError(error);
+      } else {
+        console.log("error al treer la publicacion all");
+      }
+    }
+  };
   return (
     <PortalLayout>
       <div className="perfil">
@@ -251,10 +336,9 @@ export default function Dashboard() {
               style={{ display: "none" }}
               onChange={(e) => handleProfileImageChange(e.target.files)}
             />
-            <label className="button" style={{ color: "black" }} htmlFor="profileImageInput">
+            <label className="button" style={{ color: "black" }} htmlFor="profileImageInput">+
             </label>
           </div>
-
           {downloadURL ? (
             <img
               src={downloadURL}
@@ -263,49 +347,92 @@ export default function Dashboard() {
             />
           ) : null}
         </div>
+        <div style={{ display: "flex", flexDirection: "column" }}>
+          <form className="publiText" encType="multipart/form-data" onSubmit={postPublication}>
+            <div >
+              <input type="file"
+                ref={fileInputRef}
+                name="file"
+                style={{ display: "none" }}
+                id='Publicacion' />
+              <label className='publicacion' htmlFor="Publicacion"></label>
+              <textarea
+                className="textarea"
+                name="description"
+                placeholder="Escribe nueva publicacion"
+                value={description}
+                onChange={(e) => setDescription(e.target.value)}
+                ref={descriptionTextareaRef} />
+              {error && <p>Error: {error}</p>}
+              <button
+                className="btn btn-dark my-2"
+                type="submit">Publicar</button>
+            </div>
+          </form>
+          <form className="publiText" encType="multipart/form-data" onSubmit={handleFormSubmitFalse} style={{ alignSelf: "flex-end", flexDirection: "row-reverse" }}>
+            <div>
+              <h3 style={{ color: 'black' }}>Mis publicaciones </h3>
+              <button className="btnViewAll" type="submit"></button>
+            </div>
+          </form>
 
-        <form className="publiText" encType="multipart/form-data" onSubmit={postPublication}>
-          <input type="file"
-            ref={fileInputRef}
-            name="file" 
-            style={{ display: "none" }}
-            id='Publicacion'/>
-            <label className='publicacion' htmlFor="Publicacion"></label>
-
-          <textarea
-            className="textarea"
-            name="description"
-            placeholder="Escribe nueva publicacion"
-            value={description}
-            onChange={(e) => setDescription(e.target.value)}
-            ref={descriptionTextareaRef} />
-          <h2 style={{ color: 'black', marginTop: '60px' }}>Publicaciones </h2>
-          {error && <p>Error: {error}</p>}
-          <button
-            className="btnPublication"
-            type="submit"></button>
-        </form>
+          <form className="publiText" encType="multipart/form-data" onSubmit={handleFormSubmit} style={{ display: 'flex', justifyContent: 'flex-end' }}>
+            <div style={{ alignItems: 'flex-end', margin: -130 }}>
+              <h3 style={{ color: 'black', margin: 0 }}>Ver todo</h3>
+              <button className="btnViewAll" type="submit"></button>
+            </div>
+          </form>
+        </div>
       </div>
-      
-      <ul className='collageImage'>
-     
-        {publicaciones.map((publicacion, index) => (
-          <li className='card' key={index}>
-          <div className='delete'>
-            <h2> {auth.getUser()?.name ?? ""}</h2> 
-            <button className='recycle' 
-            onClick={() => handleDeletePublication(publicacion._id)}>
-              </button>
+      {isVisible ? (
+        <ul className='collageImage'>
+          {publicaciones.map((publicacion, index) => (
+            <li className='card' key={index}>
+              <div className='delete'>
+                <h2>{auth.getUser()?.name ?? ""}</h2>
+                <button onClick={() => {
+                  modalHandler(true, publicacion.image, publicacion._id)
+                }} className='btn btn-dark'>Ver</button>
               </div>
-            <p className='decription'> {publicacion.description}</p>
-            <img className='card-image' src={publicacion.image} alt="" />
-            <button className='citas'></button><h4>AGENDA</h4>
-          </li>
-        ))}
-      </ul>
-
+              <p className='decription'>{publicacion.description}</p>
+              <img className='card-image' src={publicacion.image} alt="" />
+              <button className='citas'></button><h4>AGENDA</h4>
+            </li>
+          ))}
+        </ul>
+      ) : (
+        <ul className='collageImage'>
+          {(publicacionesUsuarios as any[]).map((publicacion, index) => (
+            <li className='card' key={index}>
+              <div className='delete'>
+                {(publicacion as any[]).map((item1, subIndex) => (<h2 className='nameUser' key={subIndex}> {item1.name ?? ""}</h2>))}
+                <button onClick={() => {
+                  modalHandler(true, publicacion.image, publicacion._id)
+                }} className='btn btn-dark'>Ver</button>
+              </div>
+              {/* Recorrer cada publicación dentro del subarray */}
+              {(publicacion as any[]).map((item, subIndex) => (
+                <div key={subIndex}>
+                  <p className='decription'> {item.description}</p>
+                  <img className='card-image' src={item.image} alt="" />
+                  <button className='citas'></button><h4>AGENDA</h4>
+                </div>
+              ))}
+            </li>
+          ))}
+        </ul>
+      )}
+      <Modal className='card' style={{ content: { width: '50%', margin: '0 auto', marginTop: '100px' } }} isOpen={modalIsOpen} onRequestClose={() => modalHandler(false, null, null)}>
+        <div >
+          <div className='card-body' style={{ display: 'flex', justifyContent: 'space-between', width: "100%" }}>
+            <button onClick={() => deleteHandler()} className='btn btn-danger'>ELIMINAR</button>
+            <button className='btn btn-danger' onClick={() => modalHandler(false, null, null)}>X</button>
+          </div>
+          <img style={{ padding: 10, width: '100%' }} src={currenImage || ''} alt="" />
+        </div>
+      </Modal>
     </PortalLayout>
-  );  
+  );
 }
